+++ odoo/songs/install/data_full.py
@@ -157,6 +157,9 @@ def location_compute_parents(ctx):
 @anthem.log
 def import_delivery_round_config(ctx):
     """ Importing delivery round config from csv"""
+    content = \
+        resource_stream(req, 'data/install/round.template.version.csv')
+    load_csv_stream(ctx, 'round.template.version', content, delimiter=',')
     content = resource_stream(req, 'data/install/delivery_template.csv')
     load_csv_stream(ctx, 'round.template', content, delimiter=',')
     content = resource_stream(req, 'data/install/delivery_itinerary.csv')