+++ HISTORY.rst
@@ -9,6 +9,8 @@ Unreleased
 **Features and Improvements**
 
 * ALCN-994: Add smartbutton for helpdesk tickets on sale and purchase
+* ALCN-999: Fix some small bugs with the module delivery_rounds
+* ALCN-999: Fix some small bugs with the module delivery_rounds and add a cron to create delivery daily plan
 
 **Bugfixes**